homework08/requirements/dev.txt
-r base.txt
django-debug-toolbar
coverage==4.5.1
html2text
